16. In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+ task

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stdint.h>
 
 /**
  * main -Entry function
@@ -7,7 +8,7 @@
  */
 int main(void)
 {
-	unsigned long long int a = 1, b = 2, s;
+	int64_t a = 1, b = 2, s;
 	int i;
 
 	printf("%llu, %llu", a, b);
